+++ tests/modeling/test_results.py
@@ -1,4 +1,3 @@
-import shutil
 from io import StringIO
 
 import numpy as np
@@ -14,9 +13,7 @@ from pharmpy.modeling import (
     calculate_pk_parameters_statistics,
     check_parameters_near_bounds,
     set_iiv_on_ruv,
-    summarize_modelfit_results,
 )
-from pharmpy.utils import TemporaryDirectoryChanger
 
 
 def test_calculate_eta_shrinkage(load_model_for_test, testdata):
@@ -134,110 +131,6 @@ k_e,median,13.319584,2.67527,2.633615
     correct.index.set_names(['parameter', 'covariates'], inplace=True)
     correct.drop('stderr', inplace=True, axis=1)
     # pd.testing.assert_frame_equal(df, correct, atol=1e-4)
-
-
-def test_summarize_modelfit_results(
-    load_model_for_test, create_model_for_test, testdata, pheno_path
-):
-    pheno = load_model_for_test(pheno_path)
-
-    summary_single = summarize_modelfit_results(pheno)
-
-    assert summary_single.loc['pheno_real']['ofv'] == 586.2760562818805
-    assert summary_single['OMEGA(1,1)_estimate'].mean() == 0.0293508
-
-    assert len(summary_single.index) == 1
-
-    mox = load_model_for_test(testdata / 'nonmem' / 'models' / 'mox1.mod')
-
-    summary_multiple = summarize_modelfit_results([pheno, mox])
-
-    assert summary_multiple.loc['mox1']['ofv'] == -624.5229577248352
-    assert summary_multiple['OMEGA(1,1)_estimate'].mean() == 0.2236304
-    assert summary_multiple['OMEGA(2,1)_estimate'].mean() == 0.395647  # One is NaN
-
-    assert len(summary_multiple.index) == 2
-    assert list(summary_multiple.index) == ['pheno_real', 'mox1']
-
-    pheno_no_res = create_model_for_test(pheno.model_code)
-    pheno_no_res.name = 'pheno_no_res'
-
-    summary_no_res = summarize_modelfit_results([pheno, pheno_no_res])
-
-    assert summary_no_res.loc['pheno_real']['ofv'] == 586.2760562818805
-    assert np.isnan(summary_no_res.loc['pheno_no_res']['ofv'])
-    assert np.all(np.isnan(summary_no_res.filter(regex='estimate$').loc['pheno_no_res']))
-
-    pheno_multest = load_model_for_test(
-        testdata
-        / 'nonmem'
-        / 'modelfit_results'
-        / 'onePROB'
-        / 'multEST'
-        / 'noSIM'
-        / 'pheno_multEST.mod'
-    )
-
-    summary_multest = summarize_modelfit_results([pheno_multest, mox])
-
-    assert len(summary_multest.index) == 2
-
-    assert not summary_multest.loc['pheno_multEST']['minimization_successful']
-
-    summary_multest_full = summarize_modelfit_results(
-        [pheno_multest, mox], include_all_estimation_steps=True
-    )
-
-    assert len(summary_multest_full.index) == 3
-    assert len(set(summary_multest_full.index.get_level_values('model'))) == 2
-    assert summary_multest_full.loc['pheno_multEST', 1]['run_type'] == 'estimation'
-    assert summary_multest_full.loc['pheno_multEST', 2]['run_type'] == 'evaluation'
-
-    assert not summary_multest_full.loc['pheno_multEST', 1]['minimization_successful']
-
-    pheno_multest_no_res = create_model_for_test(pheno_multest.model_code)
-    pheno_multest_no_res.name = 'pheno_multest_no_res'
-
-    summary_multest_full_no_res = summarize_modelfit_results(
-        [pheno_multest_no_res, mox], include_all_estimation_steps=True
-    )
-
-    assert summary_multest_full_no_res.loc['mox1', 1]['ofv'] == -624.5229577248352
-    assert np.isnan(summary_multest_full_no_res.loc['pheno_multest_no_res', 1]['ofv'])
-    estimates = summary_multest_full_no_res.loc['pheno_multest_no_res', 2].iloc[2:]
-    assert estimates.isnull().all()
-
-
-def test_summarize_modelfit_results_errors(load_model_for_test, testdata, tmp_path, pheno_path):
-    with TemporaryDirectoryChanger(tmp_path):
-        model = load_model_for_test(pheno_path)
-        shutil.copy2(testdata / 'pheno_data.csv', tmp_path)
-
-        error_path = testdata / 'nonmem' / 'errors'
-
-        shutil.copy2(testdata / 'nonmem' / 'pheno_real.mod', tmp_path / 'pheno_no_header.mod')
-        shutil.copy2(error_path / 'no_header_error.lst', tmp_path / 'pheno_no_header.lst')
-        shutil.copy2(testdata / 'nonmem' / 'pheno_real.ext', tmp_path / 'pheno_no_header.ext')
-        model_no_header = load_model_for_test('pheno_no_header.mod')
-        model_no_header.datainfo = model_no_header.datainfo.derive(path=tmp_path / 'pheno_data.csv')
-
-        shutil.copy2(testdata / 'nonmem' / 'pheno_real.mod', tmp_path / 'pheno_rounding_error.mod')
-        shutil.copy2(error_path / 'rounding_error.lst', tmp_path / 'pheno_rounding_error.lst')
-        shutil.copy2(testdata / 'nonmem' / 'pheno_real.ext', tmp_path / 'pheno_rounding_error.ext')
-        model_rounding_error = load_model_for_test('pheno_rounding_error.mod')
-        model_rounding_error.datainfo = model_rounding_error.datainfo.derive(
-            path=tmp_path / 'pheno_data.csv'
-        )
-
-        models = [model, model_no_header, model_rounding_error]
-        summary = summarize_modelfit_results(models)
-
-        assert summary.loc['pheno_real']['errors_found'] == 0
-        assert summary.loc['pheno_real']['warnings_found'] == 0
-        assert summary.loc['pheno_no_header']['errors_found'] == 2
-        assert summary.loc['pheno_no_header']['warnings_found'] == 1
-        assert summary.loc['pheno_rounding_error']['errors_found'] == 2
-        assert summary.loc['pheno_rounding_error']['warnings_found'] == 0
 
 
 def test_aic(load_model_for_test, testdata):